workers/tts/requirements.txt
piper-tts==1.3.0
huggingface_hub==0.35.0
